refactored uncategorized files
