+++ include/sealib/collection/subgraphstack.h
