+++ include/sealib/dictionary/rankstructure.h
