+++ include/sealib/dictionary/sharedrankstructure.h
