+++ src/collection/largedoublelinkedlist.cpp
@@ -1,4 +1,4 @@
-#include <sealib/largedoublelinkedlist.h>
+#include "../trail/largedoublelinkedlist.h"
 #include <iostream>
 
 Sealib::LargeDoubleLinkedList::LargeDoubleLinkedList(uint32_t size) :
