+++ src/dyck/dyckmatchingstructure.cpp
@@ -1,4 +1,4 @@
-#include <sealib/dyckmatchingstructure.h>
+#include "dyckmatchingstructure.h"
 #include <iostream>
 
 uint64_t Sealib::DyckMatchingStructure::getMatchNaive(
