+++ src/rankselect/rankstructure.cpp
@@ -1,5 +1,5 @@
-#include <sealib/localranktable.h>
-#include <sealib/rankstructure.h>
+#include "localranktable.h"
+#include <sealib/dictionary/rankstructure.h>
 #include <iostream>
 #include <cmath>
 #include <utility>
