+++ src/rankselect/sharedrankselect.cpp
@@ -1,5 +1,5 @@
-#include <sealib/sharedrankselect.h>
-#include <sealib/localselecttable.h>
+#include <sealib/dictionary/sharedrankselect.h>
+#include "localselecttable.h"
 #include <iostream>
 #include <utility>
