+++ src/trail/doublelinkedlist.h
