+++ src/trail/simpletrailstructure.h
