+++ src/trail/trailstructure.h
@@ -1,12 +1,15 @@
 #ifndef SEALIB_TRAILSTRUCTURE_H_
 #define SEALIB_TRAILSTRUCTURE_H_
 
-#include <sealib/localdycktable.h>
 #include <sealib/dictionary/rankselect.h>
 #include <sealib/dictionary/simplerankselect.h>
-#include <sealib/dyckmatchingstructure.h>
-#include <sealib/largedoublelinkedlist.h>
+#include "../dyck/dyckmatchingstructure.h"
+#include "doublelinkedlist.h"
 #include <vector>
+
+namespace Sealib {
+    class DoubleLinkedList;
+}
 
 /**
  * Space efficient TrailStructure.
